refactor(card) - 불필요한 줄바꿈 삭제
